requirements.txt
torch
torchvision
diffusers
transformers
accelerate
scipy
moviepy
ffmpeg-python
Pillow
opencv-python
pillow-heif>=0.13.0
rawpy>=0.18.0
imageio>=2.31.0
scikit-learn
flask
PyWavelets
scikit-image
yt-dlp
requests
beautifulsoup4
librosa
pybind11
tqdm
playwright
